refactor: Move block searching logic into its own class
